+++ api/controllers/OrderController.js
@@ -1,4 +1,7 @@
-const { order, meal, mealOrder } = require('../models');
+
+import moment from 'moment';
+
+import { order, meal, mealOrder } from '../models';
 
 class OrdersController {
 	getAllOrders(req, res) {
@@ -7,17 +10,25 @@ class OrdersController {
 				model: meal,
 			}],
 		}).then((orders) => {
-			res.status(200).send(orders);
+			if (orders.length) {
+				res.status(200).send(orders);
+			} else {
+				res.status(404).send({ message: 'Orders not found' });
+			}
 		});
 	}
 
-	getOrdersByCatererId(req, res) {
+	getOrdersByUserId(req, res) {
 		order.findAll({
 			include: [{
 				model: meal,
-				where: { userId: req.params.id },
-			}] }).then((orders) => {
-			res.status(200).send(orders);
+			}],
+			where: { userId: req.params.id } }).then((orders) => {
+			if (orders.length) {
+				res.status(200).send(orders);
+			} else {
+				res.status(404).send({ message: 'Orders not found' });
+			}
 		});
 	}
 
@@ -28,60 +39,79 @@ class OrdersController {
 			}],
 			where: { id: req.params.id },
 		}).then((orders) => {
-			res.status(200).send(orders);
+			if (orders) {
+				res.status(200).send(orders);
+			} else {
+				res.status(404).send({ message: 'Order not found' });
+			}
 		});
 	}
 
 	postOrder(req, res) {
-		const { userId } = req.body;
+		const { amount, createdAt } = req.body;
+		const userId = req.user.id;
+
 		const newOrder = order.build({
+			amount,
 			userId,
+			createdAt,
 		});
 
 		const newMealOrders = [];
-		const meals = JSON.parse(req.body.meals);
+		const { meals } = req.body;
 
 		meals.forEach((ml) => {
 			newMealOrders.push({
 				orderId: newOrder.id,
 				mealId: ml.mealId,
+				units: ml.units,
 			});
 		});
 
-		newOrder.save().then(() => {
+		newOrder.save().then((ord) => {
 			mealOrder.bulkCreate(newMealOrders).then(() => {
-				res.status(200).send({
+				res.status(201).send({
 					message: 'Order successfully created',
+					order: ord,
 				});
 			});
 		});
 	}
 
 	putOrder(req, res) {
-		const { userId } = req.body;
-		order.findOne({ where: { id: req.params.id } }).then(() => {
-			order.update(
-				{
-					userId,
-				},
-				{ where: { id: req.params.id }, returning: true },
-			).then((updated) => {
-				mealOrder.destroy({ where: { orderId: req.params.id } }).then(() => {
-					const newMealOrders = [];
-					const meals = JSON.parse(req.body.meals);
+		const userId = req.user.id;
+		order.findOne({ where: { id: req.params.id } }).then((ord) => {
+			if (ord) {
+				if (moment(ord.createdAt).add(60, 'minutes').isBefore(new Date())) {
+					res.status(400).send({ message: 'You cannot modify an order 60 minutes after it is placed' });
+				} else {
+					order.update(
+						{
+							userId,
+						},
+						{ where: { id: req.params.id }, returning: true },
+					).then((updated) => {
+						const update = updated[1][0];
+						mealOrder.destroy({ where: { orderId: req.params.id } }).then(() => {
+							const newMealOrders = [];
+							const { meals } = req.body;
 
-					meals.forEach((ml) => {
-						newMealOrders.push({
-							orderId: req.params.id,
-							mealId: ml.mealId,
+							meals.forEach((ml) => {
+								newMealOrders.push({
+									orderId: req.params.id,
+									mealId: ml.mealId,
+								});
+							});
+
+							mealOrder.bulkCreate(newMealOrders).then(() => {
+								res.status(200).send(update);
+							});
 						});
 					});
-
-					mealOrder.bulkCreate(newMealOrders).then(() => {
-						res.status(200).send(updated[1][0]);
-					});
-				});
-			});
+				}
+			} else {
+				res.status(404).send({ message: 'Order not found' });
+			}
 		});
 	}
 }
